skip a test on 5.8
If you "require" a module on 5.8 and it fails, it dies.  If you try to
require it again, it lives.  On 5.10, it will fail every time.

We should introduce something to track this failure, I guess, but it
will be only a half-fix.  For now, it should not actually affect the
feature fix that Kent made.