test that, with passthru disabled, `--` is not passed along
Imagine this invocation:

  appname cmdname -a one -- -z two

Common sense (I think) would lead us to believe that when the "execute"
method on the command class for "cmdname" is called, $args would
contain:

  [ '-z', 'two' ]

...but in reality, we get the leading `--`.  That's really just meant to
be an instruction to Getopt::Long, and not an argv token.  This commit
adds a test for this case.  It also sets _getopt_conf in Test::MyCmd to
[] because without that, unknown tokens are passed along, including the
`--`.