global options: always include pass_through to Getopt::Long
When interpreting:

    appname   [ cmdname --xyz ]

    ^- ARG0   ^- ARGV

Getopt::Long is run against the ARGV twice: once to pluck out global
options and command name from the start, once to process the options to
that command.

The first pass must *always* be called with $Getopt::Long::pass_through
set, because that's how "cmdname" is left in place.  Because the
arguments to Getopt::Long::Configure are an ordered list of
instructions, we can just tack "pass_through" on to the end, even if it
was already in the defaults, or follows a "no_pass_through".

This change will require Getopt::Long::Descriptive 0.116, because
previous versions do not localize their changes to Getopt::Long
configuration.